requirements.txt
absl-py==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
datasets==2.14.5
dill==0.3.7
filelock==3.12.4
Flask==2.3.3
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.9.0
huggingface-hub==0.17.2
idna==3.4
importlib-metadata==6.8.0
inflate64==0.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
multivolumefile==0.2.3
networkx==3.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.1
portalocker==2.8.2
protobuf==4.24.3
psutil==5.9.5
py7zr==0.20.6
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybcj==1.0.1
pycryptodomex==3.19.0
pyppmd==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzstd==0.15.9
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rouge-score==0.1.2
rsa==4.9
sacrebleu==2.3.1
safetensors==0.3.3
sentencepiece==0.1.99
six==1.16.0
summarizer==0.0.7
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
texttable==1.6.7
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
wrapt==1.15.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.17.0
